fs class - changed read,write,append to get size_t count instead of int count

+++ include/fs/fs.h
@@ -3,6 +3,8 @@
 
 #include "fs/path.h"
 
+#include <stdint.h>
+
 class FS {
 public:
     virtual ~FS() {};
@@ -11,9 +13,9 @@ public:
     virtual void mount(int what) = 0;
     virtual void umount() = 0;
 
-    virtual void read(Path path, int count, char* buf) = 0;
-    virtual void write(Path path, int count, char* buf) = 0;
-    virtual void append(Path path, int count, char* buf) = 0;
+    virtual void read(Path path, size_t count, char* buf) = 0;
+    virtual void write(Path path, size_t count, char* buf) = 0;
+    virtual void append(Path path, size_t count, char* buf) = 0;
 };
 
 #endif
